Revert "Add expense tracking into transaction table."
This reverts commit 356b7bc39eeb2c38e3e39c4fba883034ccfcbd59.
