+++ app/Livewire/BudgetForm.php
@@ -59,10 +59,7 @@ class BudgetForm extends Component
 
     public function calculateRemainingBalance()
     {
-        if (!$this->income) return 0;
-        $budget = Budget::where('user_id', Auth::id()) ->first();
-        if(!$budget) return $this->income;
-        return $this->income - $budget->needs_progress ?? 0 + $budget->wants_progress ?? 0 + $budget->savings_progress ?? 0;
+        $this->remaining_balance = $this->income - ($this->needs + $this->wants + $this->savings);
     }
 
     public function useAIRecommendations()
@@ -96,6 +93,7 @@ class BudgetForm extends Component
         $this->dispatch('refreshBudgetTable');
         $this->dispatch('refreshBudgetChart');
         $this->dispatch('refreshBudgetingChat');
+        $this->dispatch('refreshPercentTable');
     }
 
     public function defaultBudget()
@@ -120,6 +118,7 @@ class BudgetForm extends Component
         $this->dispatch('refreshBudgetTable');
         $this->dispatch('refreshBudgetChart');
         $this->dispatch('refreshBudgetingChat');
+        $this->dispatch('refreshPercentTable');
     }
 
     public function prioritizeDebts()
@@ -144,6 +143,7 @@ class BudgetForm extends Component
         $this->dispatch('refreshBudgetTable');
         $this->dispatch('refreshBudgetChart');
         $this->dispatch('refreshBudgetingChat');
+        $this->dispatch('refreshPercentTable');
     }
 
     public function prioritizeSavings()
@@ -168,6 +168,7 @@ class BudgetForm extends Component
         $this->dispatch('refreshBudgetTable');
         $this->dispatch('refreshBudgetChart');
         $this->dispatch('refreshBudgetingChat');
+        $this->dispatch('refreshPercentTable');
     }
 
     public function prioritizeWants()
@@ -192,6 +193,7 @@ class BudgetForm extends Component
         $this->dispatch('refreshBudgetTable');
         $this->dispatch('refreshBudgetChart');
         $this->dispatch('refreshBudgetingChat');
+        $this->dispatch('refreshPercentTable');
     }
 
     public function render()
