+++ resources/views/livewire/debt-transaction-table.blade.php
@@ -1,0 +1,3 @@
+<div>
+    {{ $this->table }}
+</div>
