+++ resources/views/rewards/index.blade.php
@@ -79,9 +79,5 @@
             </div>
         </div>
     </div>
-       <!-- User Goals Table Section -->
-    <div id="goals" class="mb-16 m-auto max-w-[21rem] sm:max-w-sm md:max-w-2xl lg:max-w-7xl lg:px-8">
-        @livewire('user-goal-table')
-    </div>
 
 </x-app-layout>
